@@ -6,7 +6,7 @@ description: Demonstrates how to use the flutter_media_info plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+    sdk: '>=3.0.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -18,10 +18,10 @@ dependencies:
   flutter:
     sdk: flutter
 
-  file_selector: ^0.9.2+2
-  file_selector_linux: ^0.9.0+1
-  file_selector_macos: ^0.9.0+3
-  file_selector_windows: ^0.9.1+4
+  file_selector: ^1.0.1
+  file_selector_linux: ^0.9.2+1
+  file_selector_macos: ^0.9.3+3
+  file_selector_windows: ^0.9.3+1
   flutter_media_info:
     # When depending on this package from a real application you should use:
     #   flutter_media_info: ^x.y.z
@@ -39,7 +39,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
