@@ -6,19 +6,19 @@ repository: https://github.com/pcpl2/flutterMediainfo
 issue_tracker: https://github.com/pcpl2/flutterMediainfo/issues
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  ffi: ^2.0.1
-  path: ^1.8.2
+  ffi: ^2.1.0
+  path: any
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
 
 flutter:
   plugin:
